opengl: add "direct" sampler
A sampler exposes the VLC input picture via a GLSL function
vlc_texture().

Until now, the input picture always came from the interop (initialized
from a picture_t). With filters, the input picture may now be the output
picture of the previous filter, already stored in a OpenGL texture.

In that case, we need a sampler able to expose this input via
vlc_texture().