contrib: lua: autodetect lua >= 5.1
When building on ArchLinux currently, the contrib was built even if
lua5.3 is installed. As at least Archlinux and Debian are exposing both
a luaX.Y.pc and lua.pc file for the latest supported version of lua,
detecting that the system version is correct through lua.pc should
also ensure that the latest compatible version is installed and used
instead.

The other checks are still needed as you can for example install lua5.2
(which will expose a lua5.2.pc file but no lua.pc file) without having
to install lua (5.3).