contrib: lua: rewrite prefix in pkg-config file
This patch enable the writing of the pkg-config file for all platforms
and enforce the correct prefix= value in the file. Previously the
pkg-config file was never installed for linux platforms so it was using
the file detection in the configure.ac, which is the last method of
detection.

If a newer lua was detected that was not in the contrib file
autodetection, it would build the lua contrib but still configure with
the newer one, which means that contrib includes will be used and then
it will be linked to the detected external lua.

In particular, on Archlinux, when using the default lua (5.3) version,
the contrib would be built but it will still link lua5.3 which has no
symbol for the luaL_register function used when header version is less
than or equal to 5.2.

With previous change, as `lua >= 5.1` is tested first, contrib will be
used first if built whatever the version of lua on the platform, so in
the worst case we only build lua from the contrib too often (like
previously) instead of failing link on some platforms too.