configure.ac: look for lua.pc first
If contrib has been built but is not detected first, contrib includes
were used but the final binaries were linked to another library.

In particular, on Archlinux, when using the default lua (5.3) version,
the contrib would be built but it would still link lua5.3 which has no
symbol for the luaL_register function used when header version is less
than or equal to 5.2.

Instead, check for lua.pc that has the most opportunity to be the
lastest on the system and is also the name of the contrib pkg-config
file when installed.