vulkan: force swapchain resize on display size change
On Windows the swapchain might not be re-created automatically, which
result in glitches when using the vk output. This enforce the creation
of the swapchain as soon as the display size is changed.

On Wayland, the size of the window is the size of the content so there
is no automatic swapchain sizing and this call is mandatory.